move publish file and update snapshot
