+++ packages/electron-builder/src/index.ts
@@ -1,6 +1,6 @@
 export { getArchSuffix, Arch, archFromString } from "builder-util"
 export { build, CliOptions, createTargets } from "./builder"
-export { publish, publishArtifactsWithOptions } from "./cli/publish"
+export { publish, publishArtifactsWithOptions } from "./publish"
 export {
   TargetConfiguration,
   Platform,
